- separated some deeply nested subcommands and grouped everything under the top-level command "feature" - added "feature" app to the pyproject.toml
