+++ git_tool/ci/subcommands/feature_add.py
@@ -14,7 +14,7 @@ from git_tool.finding_features import features_for_file_by_annotation
 app = typer.Typer(no_args_is_help=True)
 
 
-@app.command("by-add", no_args_is_help=True)
+@app.command("add", no_args_is_help=True)
 def feature_add_by_add(
     feature_names: list[str] = typer.Argument(
         None, help="List of feature names to associate with the staged files"
@@ -51,48 +51,6 @@ def feature_add_by_add(
     else:
         typer.echo("No features provided.", err=True)
 
-
-@app.command("from-staged", help="Associate staged files with features")
-def features_from_staging_area():
-    """
-    Use the staged files to add feature information.
-    """
-    feature_names = read_features_from_staged()
-    if feature_names:
-        typer.echo(f"Features in staging area: {feature_names}")
-        write_staged_featureset(feature_names)
-    else:
-        typer.echo("No features found in staging area.", err=True)
-
-
-def read_features_from_staged(type: str = "Union") -> list[str]:
-    """
-    Retrieve the features from the current staging area.
-
-    Keyword Arguments:
-        type -- Can be Union or Intersection. Describes how multiple features will be combined (default: {"Union"})
-
-    Returns:
-        List of features.
-    """
-    typer.echo("Using staged files")
-
-    staged_files = get_files_by_git_change().get("staged_files", [])
-    feature_sets = [set(get_features_for_file(f)) for f in staged_files]
-
-    if not feature_sets:
-        return []
-
-    if type == "Union":
-        combined_features = set.union(*feature_sets)
-    elif type == "Intersection":
-        combined_features = set.intersection(*feature_sets)
-    else:
-        raise ValueError(
-            "Invalid type specified. Use 'Union' or 'Intersection'."
-        )
-
-    return list(combined_features)
 
 
 def stage_files(selected_files: list[str]) -> bool:
